Task04/requirements.txt
streamlit
langchain
huggingface_hub
sentence-transformers
faiss-cpu
python-dotenv
